Update AutonSquare.java
Updated the AutonSquare and tested the teleop drives. A lot of spin out on the teleop from the back right motor when going full speed ahead. The Auton P_DRIVE_GAIN was taken from 0.03 to 0.00 because every straight take off in Auton was spinning out consistently even after drive was lowered from 0.4 to 0.2. Near perfect after that adjustment in Auton