introducing driver control
I stole some of Mr. Vanscoy's drive code and modified it to fit my purposes. I didn't really know what I was doing but hopefully this works. The code is supposed to go right 36 inches when the gamepad1 right bumper is pressed, and left 36 inches when the gamepad1 left bumper is pressed.