Fixed strafe/bumper code
I fixed the gamepad1-bumper-code that strafes 36 inches either left or right depending on which bumper on gamepad1 is pressed. I do not have working code for the left bumper at the moment, but, along with a few other modifications to these particular blocks of code I'm planning, are on their way.