comments and zero power for floor position
I added some of the last comments I think this class needs to clarify what it says at a first glance or to those unacquainted with FTC Java code. When I was editing earlier, I must've accidentally taken out the code that set the floor position's power to zero to prevent motor burnout. I put that command back where it belongs, but not after some issues at a driver practice.