Sped up lift and telemetry experiment
Per the drivers' request, I sped up the lift and attempted to experiment with the joysticks to make them less sensitive, but could not do so as the Driver Control Hub was having issues. I expect to work on this later after school, if the Driver Control Hub is operational.