Changes to the gamepad1 bumper
I made some more changes in hope of fixing the gamepad1 bumper strafe, only to come away realizing I had more errors than I initially thought. Not only is it not going the correct distance, but after the command is executed, the almost half of the controls cease working. I hope to bring my dad in tomorrow to assist me with this.